requirements.txt
flask
google-generativeai
flask-sqlalchemy
flask-login
flask-bcrypt
flask-cors
python-dotenv
Authlib
requests
